requirements.txt
streamlit==1.28.0
streamlit-folium==0.15.0
pandas==2.1.1
folium==0.14.0
requests==2.31.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
openpyxl==3.1.2